% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA_convert.R
\name{ERA_convert}
\alias{ERA_convert}
\title{Convert codified Enhanced Reporting of Aboriginality (ERA) into plain English}
\usage{
ERA_convert(data)
}
\arguments{
\item{data}{Specify the dataframe column or vector to convert into plain English}
}
\value{
Equivalent plain English value of codified ERA values
}
\description{
Converts codified data (0, 1, 8, 9) into their respective plain English outputs.
}
\examples{
ERA_convert(APDC$ERA)
ERA_convert(EDDC$ERA)
ERA_convert(RBDM$ERA)

}
