% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpreter_convert.R
\name{Interpreter_convert}
\alias{Interpreter_convert}
\title{Convert codified NEED_INTERPRETER_SERVICE values into plain English}
\usage{
Interpreter_convert(data)
}
\arguments{
\item{data}{The dataframe column or vector to convert into plain English}
}
\value{
Converts codified NEED_INTERPRETER_SERVICE values into their respective plain English equivalent
}
\description{
Converts codified NEED_INTERPRETER_SERVICE values (0, 1) into their respective plain English outputs.
}
\details{
Converts codified data into their respective outputs that denote the patient's need for an interpreter
}
\examples{
Interpreter_convert(APDR$NEED_INTERPRETER_SERVICE)
}
