% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACC_convert.R
\name{SACC_convert}
\alias{SACC_convert}
\title{Convert SACC codes into English}
\usage{
SACC_convert(
  data,
  Country.region = TRUE,
  SACC_Major_group = TRUE,
  Region = TRUE
)
}
\arguments{
\item{data}{The dataframe column, or vector containing SACC codes to be converted to English.}
}
\value{
Plain English equivalents of SACC codified country names.
}
\description{
Convert SACC codes into English
}
\examples{
SACC_convert(EDDC$country_of_birth)
SACC_convert(APDC$country_of_birth_sacc)
}
